\name{datetimeutils-package}
\alias{datetimeutils-package}
\alias{datetimeutils}
\docType{package}
\title{
\packageTitle{datetimeutils}
}
\description{
\packageDescription{datetimeutils}
}
\details{
Helper functions for dealing with times and dates. 
}
\author{
  Enrico Schumann

  Maintainer: \packageMaintainer{datetimeutils}
}
\references{
  Ripley Hornik
}
\keyword{ package }
%% \seealso{
%% ~~ Optional links to other man pages, e.g. ~~
%% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%% }
%% \examples{
%% ~~ simple examples of the most important functions ~~
%% }
