\name{Easter}
\alias{Easter}
\docType{data}
\title{
  Easter Sundays
}
\description{

  A vector of Easter sundays in the Gregorian Calendar.

}
\usage{
Easter
}
\format{
  a named vector of dates
}
\details{
  A named vector of dates.
}
\source{
  \url{https://www.census.gov/data/software/x13as/genhol/easter-dates.html}
}
\references{
  \url{https://web.archive.org/web/20250000000000*/https://www.census.gov/data/software/x13as/genhol/easter-dates.html}
}
\examples{
Easter["1970"]
}
\keyword{datasets}
