\name{datetimeutils-package}
\alias{datetimeutils-package}
\alias{datetimeutils}
\docType{package}
\title{
\packageTitle{datetimeutils}
}
\description{
\packageDescription{datetimeutils}
}
\details{
  Helper functions for dealing with times and dates. 
}
\author{
  Enrico Schumann

  Maintainer: \packageMaintainer{datetimeutils}
}
\references{
  B.D. Ripley and K. Hornik. \emph{Date-Time Classes}. R-News,
  \strong{1}(2):8--12, 2001.
}
\keyword{ package }
\seealso{
  \code{\link{DateTimeClasses}}, \code{\link{Dates}}
}
